Conflicts:
	hoomd/CMakeLists.txt
	hoomd/CommunicatorGPU.cc
	hoomd/GPUVector.h
	hoomd/RandomNumbers.h
	hoomd/extern/neighbor
	hoomd/hpmc/ComputeFreeVolumeGPU.cuh
	hoomd/hpmc/IntegratorHPMCMono.h
	hoomd/hpmc/IntegratorHPMCMonoGPU.cu
	hoomd/hpmc/IntegratorHPMCMonoGPU.cuh
	hoomd/hpmc/IntegratorHPMCMonoGPU.h
	hoomd/hpmc/MAP3D.h
	hoomd/hpmc/ShapeConvexPolyhedron.h
	hoomd/hpmc/ShapeFacetedEllipsoid.h
	hoomd/hpmc/ShapePolyhedron.h
	hoomd/hpmc/ShapeSphere.h
	hoomd/hpmc/ShapeSpheropolyhedron.h
	hoomd/hpmc/ShapeUnion.h
	hoomd/hpmc/UpdaterClusters.h
	hoomd/hpmc/integrate.py
	hoomd/hpmc/test/test_convex_polyhedron.cc
	hoomd/hpmc/test/test_sphere.cc
	hoomd/hpmc/update.py
	hoomd/jit/KaleidoscopeJIT.h
	hoomd/jit/PatchEnergyJITGPU.cc
	hoomd/jit/PatchEnergyJITUnionGPU.cc

+++ hoomd/CachedAllocator.h
@@ -18,6 +18,7 @@
 
 #include <map>
 #include <cassert>
+#include <stdexcept>
 
 //! Need to define an error checking macro that can be used in .cu files
 #define CHECK_CUDA() \
@@ -76,7 +77,7 @@ class __attribute__((visibility("default"))) CachedAllocator
          * \returns a pointer to the allocated buffer
          */
         template<typename T>
-        T *getTemporaryBuffer(unsigned int num_elements);
+        T *getTemporaryBuffer(size_t num_elements);
 
         // Specifically allocate a char* buffer
         char *allocate(std::ptrdiff_t num_bytes)
@@ -105,8 +106,8 @@ class __attribute__((visibility("default"))) CachedAllocator
 
         bool m_managed;  //! True if we use unified memory
 
-        unsigned int m_num_bytes_tot;
-        unsigned int m_max_cached_bytes;
+        size_t m_num_bytes_tot;
+        size_t m_max_cached_bytes;
         float m_cache_reltol;
 
         free_blocks_type m_free_blocks;
@@ -142,7 +143,7 @@ class ScopedAllocation
     {
     public:
         //! Copy constructor
-        ScopedAllocation(CachedAllocator& alloc, unsigned int num_elements);
+        ScopedAllocation(CachedAllocator& alloc, size_t num_elements);
 
         //! Destructor
         ~ScopedAllocation();
@@ -163,7 +164,7 @@ class ScopedAllocation
     };
 
 template<typename T>
-T* CachedAllocator::getTemporaryBuffer(unsigned int num_elements)
+T* CachedAllocator::getTemporaryBuffer(size_t num_elements)
     {
     std::ptrdiff_t num_bytes = sizeof(T)*num_elements;
     char *result = 0;
@@ -171,7 +172,7 @@ T* CachedAllocator::getTemporaryBuffer(unsigned int num_elements)
     // short-cut to avoid storing duplicate NULL ptrs in the map
     if (!num_bytes) return (T*)NULL;
 
-    unsigned int num_allocated_bytes = num_bytes;
+    size_t num_allocated_bytes = num_bytes;
 
     // search the cache for a free block
     free_blocks_type::iterator free_block = m_free_blocks.lower_bound(num_bytes);
@@ -232,7 +233,7 @@ T* CachedAllocator::getTemporaryBuffer(unsigned int num_elements)
 
 //! Constructor
 template<typename T>
-ScopedAllocation<T>::ScopedAllocation(CachedAllocator& alloc, unsigned int num_elements)
+ScopedAllocation<T>::ScopedAllocation(CachedAllocator& alloc, size_t num_elements)
     : m_alloc(alloc)
     {
     data = m_alloc.getTemporaryBuffer<T>(num_elements);
