+++ hoomd/md/NeighborListGPUBinned.cu
@@ -335,7 +335,7 @@ inline void launcher(unsigned int *d_nlist,
     {
     // shared memory = r_listsq + Nmax + stuff needed for neighborlist (computed below)
     Index2D typpair_idx(ntypes);
-    unsigned int shared_size = sizeof(Scalar)*typpair_idx.getNumElements() + sizeof(unsigned int)*ntypes;
+    unsigned int shared_size = (unsigned int)(sizeof(Scalar)*typpair_idx.getNumElements() + sizeof(unsigned int)*ntypes);
 
     unsigned int offset = range.first;
     unsigned int nwork = range.second - range.first;
