+++ hoomd/md/TableDihedralForceGPU.cu
@@ -39,7 +39,7 @@
 */
 __global__ void gpu_compute_table_dihedral_forces_kernel(Scalar4* d_force,
                                      Scalar* d_virial,
-                                     const unsigned int virial_pitch,
+                                     const size_t virial_pitch,
                                      const unsigned int N,
                                      const Scalar4 *device_pos,
                                      const BoxDim box,
@@ -300,7 +300,7 @@ __global__ void gpu_compute_table_dihedral_forces_kernel(Scalar4* d_force,
 */
 hipError_t gpu_compute_table_dihedral_forces(Scalar4* d_force,
                                      Scalar* d_virial,
-                                     const unsigned int virial_pitch,
+                                     const size_t virial_pitch,
                                      const unsigned int N,
                                      const Scalar4 *device_pos,
                                      const BoxDim &box,
